Fixed small problems with types

+++ StyleGuide/lucky_tickets.c
@@ -4,7 +4,7 @@ unsigned int countLuckyTickets(void)
     for (int digit_1 = 0; digit_1 < 10; ++digit_1) {
         for (int digit_2 = 0; digit_2 < 10; ++digit_2) {
             for (int digit_3 = 0; digit_3 < 10; ++digit_3) {
-                unsigned short amount = digit_1 + digit_2 + digit_3;
+                unsigned int amount = digit_1 + digit_2 + digit_3;
                 variants[amount] += 1;
             }
         }
